Add new secureboot file for tests
